+++ core-bundle/src/Resources/contao/drivers/DC_Table.php
@@ -129,7 +129,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 		if (isset($_GET['clipboard']))
 		{
 			$objSession->set('CLIPBOARD', array());
-			$this->redirect($this->getReferer());
+			$this->redirect($this->getReferer(false, $strTable));
 		}
 
 		// Check whether the table is defined
@@ -183,7 +183,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 					$this->redirect(str_replace('act=select', 'act=copyAll', \Environment::get('request')));
 				}
 
-				$this->redirect($this->getReferer());
+				$this->redirect($this->getReferer(false, $strTable));
 			}
 		}
 
@@ -720,7 +720,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 			}
 		}
 
-		$this->redirect($this->getReferer());
+		$this->redirect($this->getReferer(false, $this->strTable));
 	}
 
 
@@ -743,7 +743,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 		// ID and PID are mandatory
 		if (!$this->intId || !strlen(\Input::get('pid')))
 		{
-			$this->redirect($this->getReferer());
+			$this->redirect($this->getReferer(false, $this->strTable));
 		}
 
 		// Get the new position
@@ -816,7 +816,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 
 		if (!$blnDoNotRedirect)
 		{
-			$this->redirect($this->getReferer());
+			$this->redirect($this->getReferer(false, $this->strTable));
 		}
 	}
 
@@ -849,7 +849,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 			}
 		}
 
-		$this->redirect($this->getReferer());
+		$this->redirect($this->getReferer(false, $this->strTable));
 	}
 
 
@@ -871,7 +871,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 
 		if (!$this->intId)
 		{
-			$this->redirect($this->getReferer());
+			$this->redirect($this->getReferer(false, $this->strTable));
 		}
 
 		/** @var SessionInterface $objSession */
@@ -1038,7 +1038,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 
 		if (!$blnDoNotRedirect)
 		{
-			$this->redirect($this->getReferer());
+			$this->redirect($this->getReferer(false, $this->strTable));
 		}
 
 		return false;
@@ -1201,7 +1201,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 			}
 		}
 
-		$this->redirect($this->getReferer());
+		$this->redirect($this->getReferer(false, $this->strTable));
 	}
 
 
@@ -1471,7 +1471,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 
 		if (!$this->intId)
 		{
-			$this->redirect($this->getReferer());
+			$this->redirect($this->getReferer(false, $this->strTable));
 		}
 
 		$delete = array();
@@ -1483,7 +1483,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 						   ->limit(1)
 						   ->execute($this->intId);
 
-			$this->redirect($this->getReferer());
+			$this->redirect($this->getReferer(false, $this->strTable));
 		}
 
 		// If there is a PID field but no parent table
@@ -1580,7 +1580,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 
 		if (!$blnDoNotRedirect)
 		{
-			$this->redirect($this->getReferer());
+			$this->redirect($this->getReferer(false, $this->strTable));
 		}
 	}
 
@@ -1612,7 +1612,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 			}
 		}
 
-		$this->redirect($this->getReferer());
+		$this->redirect($this->getReferer(false, $this->strTable));
 	}
 
 
@@ -1682,7 +1682,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 		// Check whether there is a record
 		if ($objRecords->numRows < 1)
 		{
-			$this->redirect($this->getReferer());
+			$this->redirect($this->getReferer(false, $this->strTable));
 		}
 
 		$error = false;
@@ -1691,7 +1691,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 
 		if (!is_array($data))
 		{
-			$this->redirect($this->getReferer());
+			$this->redirect($this->getReferer(false, $this->strTable));
 		}
 
 		$arrFields = array();
@@ -1752,7 +1752,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 						   ->execute($this->intId);
 		}
 
-		$this->redirect($this->getReferer());
+		$this->redirect($this->getReferer(false, $this->strTable));
 	}
 
 
@@ -1780,7 +1780,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 			}
 		}
 
-		$this->redirect($this->getReferer());
+		$this->redirect($this->getReferer(false, $this->strTable));
 	}
 
 
@@ -2092,7 +2092,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 		$return = $version . \Message::generate() . ($this->noReload ? '
 <p class="tl_error">'.$GLOBALS['TL_LANG']['ERR']['general'].'</p>' : '') . '
 <div id="tl_buttons">' . (\Input::get('nb') ? '&nbsp;' : '
-<a href="'.$this->getReferer(true).'" class="header_back" title="'.\StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['backBTTitle']).'" accesskey="b" onclick="Backend.getScrollOffset()">'.$GLOBALS['TL_LANG']['MSC']['backBT'].'</a>') . '
+<a href="'.$this->getReferer(true, $this->strTable).'" class="header_back" title="'.\StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['backBTTitle']).'" accesskey="b" onclick="Backend.getScrollOffset()">'.$GLOBALS['TL_LANG']['MSC']['backBT'].'</a>') . '
 </div>
 <form action="'.ampersand(\Environment::get('request'), true).'" id="'.$this->strTable.'" class="tl_form tl_edit_form" method="post" enctype="' . ($this->blnUploadable ? 'multipart/form-data' : 'application/x-www-form-urlencoded') . '"'.(!empty($this->onsubmit) ? ' onsubmit="'.implode(' ', $this->onsubmit).'"' : '').'>
 <div class="tl_formbody_edit">
@@ -2187,7 +2187,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 				\Message::reset();
 				\System::setCookie('BE_PAGE_OFFSET', 0, 0);
 
-				$this->redirect($this->getReferer());
+				$this->redirect($this->getReferer(false, $this->strTable));
 			}
 			elseif (isset($_POST['saveNedit']))
 			{
@@ -2621,7 +2621,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 				if (isset($_POST['saveNclose']))
 				{
 					\System::setCookie('BE_PAGE_OFFSET', 0, 0);
-					$this->redirect($this->getReferer());
+					$this->redirect($this->getReferer(false, $this->strTable));
 				}
 
 				$this->reload();
@@ -2693,7 +2693,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 		// Return
 		return '
 <div id="tl_buttons">
-<a href="'.$this->getReferer(true).'" class="header_back" title="'.\StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['backBTTitle']).'" accesskey="b" onclick="Backend.getScrollOffset()">'.$GLOBALS['TL_LANG']['MSC']['backBT'].'</a>
+<a href="'.$this->getReferer(true, $this->strTable).'" class="header_back" title="'.\StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['backBTTitle']).'" accesskey="b" onclick="Backend.getScrollOffset()">'.$GLOBALS['TL_LANG']['MSC']['backBT'].'</a>
 </div>'.$return;
 	}
 
@@ -2941,7 +2941,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 				if (isset($_POST['saveNclose']))
 				{
 					\System::setCookie('BE_PAGE_OFFSET', 0, 0);
-					$this->redirect($this->getReferer());
+					$this->redirect($this->getReferer(false, $this->strTable));
 				}
 
 				$this->reload();
@@ -3012,7 +3012,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 		// Return
 		return '
 <div id="tl_buttons">
-<a href="'.$this->getReferer(true).'" class="header_back" title="'.\StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['backBTTitle']).'" accesskey="b" onclick="Backend.getScrollOffset()">'.$GLOBALS['TL_LANG']['MSC']['backBT'].'</a>
+<a href="'.$this->getReferer(true, $this->strTable).'" class="header_back" title="'.\StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['backBTTitle']).'" accesskey="b" onclick="Backend.getScrollOffset()">'.$GLOBALS['TL_LANG']['MSC']['backBT'].'</a>
 </div>'.$return;
 	}
 
@@ -3471,7 +3471,7 @@ class DC_Table extends \DataContainer implements \listable, \editable
 		// Begin buttons container
 		$return = \Message::generate() . '
 <div id="tl_buttons">'.((\Input::get('act') == 'select') ? '
-<a href="'.$this->getReferer(true).'" class="header_back" title="'.\StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['backBTTitle']).'" accesskey="b" onclick="Backend.getScrollOffset()">'.$GLOBALS['TL_LANG']['MSC']['backBT'].'</a> ' : (isset($GLOBALS['TL_DCA'][$this->strTable]['config']['backlink']) ? '
+<a href="'.$this->getReferer(true, $this->strTable).'" class="header_back" title="'.\StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['backBTTitle']).'" accesskey="b" onclick="Backend.getScrollOffset()">'.$GLOBALS['TL_LANG']['MSC']['backBT'].'</a> ' : (isset($GLOBALS['TL_DCA'][$this->strTable]['config']['backlink']) ? '
 <a href="contao/main.php?'.$GLOBALS['TL_DCA'][$this->strTable]['config']['backlink'].'" class="header_back" title="'.\StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['backBTTitle']).'" accesskey="b" onclick="Backend.getScrollOffset()">'.$GLOBALS['TL_LANG']['MSC']['backBT'].'</a> ' : '')) . ((\Input::get('act') != 'select' && !$blnClipboard && !$GLOBALS['TL_DCA'][$this->strTable]['config']['closed'] && !$GLOBALS['TL_DCA'][$this->strTable]['config']['notCreatable']) ? '
 <a href="'.$this->addToUrl('act=paste&amp;mode=create').'" class="header_new" title="'.\StringUtil::specialchars($GLOBALS['TL_LANG'][$this->strTable]['new'][1]).'" accesskey="n" onclick="Backend.getScrollOffset()">'.$GLOBALS['TL_LANG'][$this->strTable]['new'][0].'</a> ' : '') . ($blnClipboard ? '
 <a href="'.$this->addToUrl('clipboard=1').'" class="header_clipboard" title="'.\StringUtil::specialchars($GLOBALS['TL_LANG']['MSC']['clearClipboard']).'" accesskey="x">'.$GLOBALS['TL_LANG']['MSC']['clearClipboard'].'</a> ' : $this->generateGlobalButtons()) . '