+++ core-bundle/src/Resources/contao/library/Contao/System.php
@@ -293,7 +293,7 @@ abstract class System
 		}
 
 		// Use a specific referer
-		if ($strTable != '' && isset($session[$strTable]) && \Input::get('act') != 'select')
+		if ($strTable != '' && isset($session[$strTable]))
 		{
 			$session['current'] = $session[$strTable];
 		}